refactor: 라우터 설정을 main.tsx에서 routes/index.tsx로 이동
